# apps/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/branches/{branchId}/orders": {
      "post": {
        "operationId": "OrderController_createOrder",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create order",
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrderController_findAllOrders",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tableId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DINE_IN",
                "QR_SELF_ORDER",
                "POS",
                "SELF_SERVICE_KIOSK",
                "TAKEAWAY",
                "DELIVERY",
                "MOBILE_APP",
                "WEB"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "CONFIRMED",
                "PREPARING",
                "READY",
                "SERVED",
                "COMPLETED",
                "CANCELLED",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List orders",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrderController_findOneOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get order details",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/accept": {
      "patch": {
        "operationId": "OrderController_acceptOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Accept order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/reject": {
      "patch": {
        "operationId": "OrderController_rejectOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Reject order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "operationId": "OrderController_updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update order status",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrderController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cancel order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/order-items/{id}/status": {
      "patch": {
        "operationId": "OrderController_updateOrderItemStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderItemStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update order item status",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/status-log": {
      "get": {
        "operationId": "OrderController_getOrderStatusLog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get order status change history",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "milky",
    "description": "The milky API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "MetaDataDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "total": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
      "GenericResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDataDto"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY",
              "SERVED",
              "COMPLETED",
              "CANCELLED",
              "REJECTED"
            ]
          },
          "totalAmount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "branchId",
          "status",
          "totalAmount"
        ]
      },
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "productNameSnapshot": {
            "type": "string"
          },
          "unitPrice": {
            "type": "number"
          },
          "quantity": {
            "type": "number",
            "minimum": 1
          },
          "subtotal": {
            "type": "number"
          },
          "stationType": {
            "type": "string",
            "enum": [
              "KITCHEN",
              "BAR"
            ]
          }
        },
        "required": [
          "productId",
          "productNameSnapshot",
          "unitPrice",
          "quantity",
          "subtotal",
          "stationType"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "tableId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "tableSessionId": {
            "type": "string"
          },
          "guestSessionId": {
            "type": "string"
          },
          "channel": {
            "type": "string",
            "enum": [
              "DINE_IN",
              "QR_SELF_ORDER",
              "POS",
              "SELF_SERVICE_KIOSK",
              "TAKEAWAY",
              "DELIVERY",
              "MOBILE_APP",
              "WEB"
            ]
          },
          "subtotal": {
            "type": "number"
          },
          "discountAmount": {
            "type": "number",
            "default": 0
          },
          "taxAmount": {
            "type": "number",
            "default": 0
          },
          "serviceCharge": {
            "type": "number",
            "default": 0
          },
          "tipAmount": {
            "type": "number",
            "default": 0
          },
          "totalAmount": {
            "type": "number"
          },
          "specialInstructions": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "subtotal",
          "totalAmount"
        ]
      },
      "RejectOrderDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY",
              "SERVED",
              "COMPLETED",
              "CANCELLED",
              "REJECTED"
            ]
          },
          "rejectionReasonCode": {
            "type": "string",
            "enum": [
              "ITEM_UNAVAILABLE",
              "KITCHEN_CAPACITY",
              "MODIFICATION_IMPOSSIBLE",
              "ALLERGY_CONCERN",
              "RESTAURANT_CLOSING",
              "OTHER"
            ]
          },
          "rejectionReason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdateOrderItemStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PREPARING",
              "READY",
              "SERVED",
              "UNAVAILABLE",
              "CANCELLED"
            ]
          },
          "unavailableReason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}